* commit '8a30a24c28343863680930ff7c602ecd9d65e34b':
  SAI-1186 Fixed an exception that could occur when harvesting zero documents
  SAI-1186 Implemented redirects for web requests
  SAI-1186 Work-around_ Added a flag for removing cache files prior to cache initialization. This can be used for big harvesters to improve initialization and harvesting performance drastically.
  SAI-1186 Added some minor fixes for the AbstractCache
  SAI-1186 Added missing constants
  SAI-1186 Fixed many FileUtils bugs and added more logging
  SAI-1186 Incremented Bugfix version
  SAI-974 Removed Snapshot versions
  SAI-991 Updated json libraries version
  SAI-953 Fixed a bug where old cache instances were kept after resets, causing multiple sets of cached documents to be saved and submitted after resetting the service

+++ src/main/java/de/gerdiproject/harvest/utils/cache/HarvesterCache.java
@@ -20,6 +20,8 @@ import java.io.File;
 import de.gerdiproject.harvest.IDocument;
 import de.gerdiproject.harvest.MainContext;
 import de.gerdiproject.harvest.application.events.ContextDestroyedEvent;
+import de.gerdiproject.harvest.config.Configuration;
+import de.gerdiproject.harvest.config.constants.ConfigurationConstants;
 import de.gerdiproject.harvest.event.EventSystem;
 import de.gerdiproject.harvest.harvester.events.GetProviderNameEvent;
 import de.gerdiproject.harvest.save.HarvestSaver;
@@ -149,9 +151,16 @@ public class HarvesterCache
      */
     public void applyChanges(boolean isSuccessful, boolean isAborted)
     {
+        if (changesCache.size() == 0)
+            return;
+
         changesCache.applyChanges();
 
-        if (isSuccessful && !isAborted) {
+        // work-around SAI-1186: delete old cache at the beginning if the corresponding flag is false
+        Configuration config = MainContext.getConfiguration();
+        boolean persistCache = config.getParameterValue(ConfigurationConstants.PERSIST_CACHE, Boolean.class);
+
+        if (isSuccessful && !isAborted && persistCache) {
             changesCache.forEach((String documentId, DataCiteJson document) -> {
                 if (document == null)
                     versionsCache.removeFile(documentId);
@@ -173,6 +182,14 @@ public class HarvesterCache
      */
     public void init(String hash, int harvestStartIndex, int harvestEndIndex)
     {
+        // work-around SAI-1186: delete old cache at the beginning if the corresponding flag is false
+        Configuration config = MainContext.getConfiguration();
+
+        if (!config.getParameterValue(ConfigurationConstants.PERSIST_CACHE, Boolean.class)) {
+            File stableFolder = new File(versionsCache.stableFolderPath).getParentFile();
+            FileUtils.deleteFile(stableFolder);
+        }
+
         final String harvesterHash = hash == null
                                      ? null
                                      : HashGenerator.instance().getShaHash(hash + harvestStartIndex + harvestEndIndex);
